requirements.txt
backports-abc==0.5
certifi==2016.9.26
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
matlabengineforpython===R2016b
nose==1.3.7
psycopg2==2.6.2
requests==2.12.3
singledispatch==3.4.0.3
tornado==4.4.2
virtualenv==15.1.0
Werkzeug==0.11.11
